pyproject.toml
[tool.poetry]
name = "review-ai"
version = "0.1.0"
description = ""
authors = ["Felipe G. Martins"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kaggle = "^1.6.12"
polars = "^0.20.25"
duckdb = "^0.10.2"
google-generativeai = "^0.5.2"
pre-commit = "^3.7.1"
black = "^24.4.2"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
python-docx = "^1.1.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
